v2.4.14.1 -> v2.4.14.2
  - Ivan Kokshaysky: fix alpha dec_and_lock with modules, for alpha config entry
  - Kai Germaschewski: ISDN updates
  - Jeff Garzik: network driver updates, sysv fs update
  - Kai Mäkisara: SCSI tape update
  - Alan Cox: large drivers merge
  - Nikita Danilov: reiserfs procfs information
  - Andrew Morton: ext3 merge
  - Christoph Hellwig: vxfs livelock fix
  - Trond Myklebust: NFS updates
  - Jens Axboe: cpqarray + cciss dequeue fix
  - Tim Waugh: parport_serial base_baud setting
  - Matthew Dharm: usb-storage Freecom driver fixes
  - Dave McCracken: wait4() thread group race fix

+++ include/asm-arm/arch-sa1100/hardware.h
@@ -62,7 +62,7 @@
 #ifndef __ASSEMBLY__
 
 #if 0
-# define __REG(x)	(*((volatile unsigned long *)io_p2v(x)))
+# define __REG(x)	(*((volatile u32 *)io_p2v(x)))
 #else
 /*
  * This __REG() version gives the same results as the one above,  except
@@ -70,12 +70,12 @@
  * assembly code for access to contigous registers.  It's a shame that gcc
  * doesn't guess this by itself.
  */
-typedef struct { volatile unsigned long offset[4096]; } __regbase;
+typedef struct { volatile u32 offset[4096]; } __regbase;
 # define __REGP(x)	((__regbase *)((x)&~4095))->offset[((x)&4095)>>2]
 # define __REG(x)	__REGP(io_p2v(x))
 #endif
 
-# define __PREG(x)	(io_v2p((unsigned long)&(x)))
+# define __PREG(x)	(io_v2p((u32)&(x)))
 
 #else
